+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 1.0.0 - 1.0.4,
 
 - First initiate version.
+
+## 1.0.5
+
+- add paste option from clipboard on long press inside field.
+- some bug fixed.
+- add feature like custom valid regex.
+- text customization of any static string inside the plugin.